[Dico] Fix Math.random off by one error.
The ending parameter is excluded, but 255 is part of the rgb color range
so I bumped it up to 256.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Math/random?redirectlocale=en-US&redirectslug=JavaScript%2FReference%2FGlobal_Objects%2FMath%2Frandom

Hi Grace,

I'm sorry for being a jerk last night.  You did say check it out.
I was in product building mood, not friend mood.  I'm sorry for not
being able to break away for ya.  You deserve better.

The random colors is pretty sweet.

I also love your comment "Include the required Google branding" which...
is indefinitely a TODO.

Cheers,
Taylor